Fix: Solved some PR comments.
